correct formula

+++ app/src/main/java/com/bitmark/autonomy/util/modelview/AreaModelView.kt
@@ -28,16 +28,16 @@ data class AreaModelView(
     }
 }
 
-fun toDrawableRes(value: Float) = when {
-    value == 0f -> R.drawable.ic_circle_mine_shaft_2
-    value < 34f -> R.drawable.ic_circle_red
-    value < 67f -> R.drawable.ic_circle_yellow
+fun toDrawableRes(value: Int) = when {
+    value == 0 -> R.drawable.ic_circle_mine_shaft_2
+    value < 34 -> R.drawable.ic_circle_red
+    value < 67 -> R.drawable.ic_circle_yellow
     else -> R.drawable.ic_circle_green
 }
 
-fun toColorRes(value: Float) = when {
-    value == 0f -> R.color.mine_shaft_2
-    value < 34f -> R.color.persian_red
-    value < 67f -> R.color.gold_tip
+fun toColorRes(value: Int) = when {
+    value == 0 -> R.color.mine_shaft_2
+    value < 34 -> R.color.persian_red
+    value < 67 -> R.color.gold_tip
     else -> R.color.apple
 }
